refactor(src/game): review changes
