+++ packages/client/src/game/entities/Explosion/Explosion.ts
@@ -1,5 +1,5 @@
-import { Color } from '../../services/View/data/colors';
-import { spriteCoordinates } from '../../services/View/data/spriteCoordinates';
+import { Color } from '../../services/View/colors';
+import { spriteCoordinates } from '../../services/View/spriteCoordinates';
 import { type Projectile, type Tank, Entity } from '../';
 import { Direction, EntityEvent } from '../Entity/typings';
 import { type ExplosionVariant } from './typings';
