+++ packages/client/src/game/entities/Tank/Tank.ts
@@ -1,9 +1,9 @@
-import { Color } from '../../services/View/data/colors';
-import { spriteCoordinates } from '../../services/View/data/spriteCoordinates';
+import { Color } from '../../services/View/colors';
+import { spriteCoordinates } from '../../services/View/spriteCoordinates';
 import { EntityDynamic, Projectile } from '../';
 import { Direction, EntityEvent } from '../Entity/typings';
+import { Speed } from '../EntityDynamic/data';
 import { type EntityDynamicSettings } from '../EntityDynamic/typings';
-import { Speed } from './data';
 
 export class Tank extends EntityDynamic {
   width = 4;
