+++ packages/client/src/game/entities/Terrain/Terrain.ts
@@ -1,5 +1,5 @@
-import { Color } from '../../services/View/data/colors';
-import { spriteCoordinates } from '../../services/View/data/spriteCoordinates';
+import { Color } from '../../services/View/colors';
+import { spriteCoordinates } from '../../services/View/spriteCoordinates';
 import { Entity } from '../';
 import { type EntitySettings, EntityEvent } from '../Entity/typings';
 import { type TerrainVariant } from './typings';
