+++ packages/client/src/game/services/View/spriteCoordinates.ts
