+++ packages/client/src/game/ui/screens/GameScreens/LevelSelectorScreen.ts
@@ -1,4 +1,4 @@
-import { Color } from '../../../services/View/data/colors';
+import { Color } from '../../../services/View/colors';
 import { Screen } from '../Screen';
 import { type LevelSelectorScreenProps } from './typings';
