+++ packages/client/src/game/ui/screens/UIScreens/LoadingScreen.ts
@@ -1,4 +1,4 @@
-import { Color } from '../../../services/View/data/colors';
+import { Color } from '../../../services/View/colors';
 import { Screen } from '../Screen';
 
 export class LoadingScreen extends Screen {
