+++ packages/client/src/game/ui/screens/UIScreens/MainMenuScreen.ts
@@ -1,6 +1,6 @@
 import { resources } from '../../../services';
-import { Color } from '../../../services/View/data/colors';
-import { spriteCoordinates } from '../../../services/View/data/spriteCoordinates';
+import { Color } from '../../../services/View/colors';
+import { spriteCoordinates } from '../../../services/View/spriteCoordinates';
 import { Screen } from '../Screen';
 import { MainMenuState } from './data';
